Widen hid-io joystick axes to 16-bit for fader resolution
The slide faders carry an absolute position whose useful resolution --
especially across the taper-compressed top of travel -- is far finer
than the 8-bit axes allowed (the top third collapsed to ~10 codes).
Widen the report body axes to int16_t and bump the descriptor to
REPORT_SIZE 16 / LOGICAL_MIN16/MAX16 so the full ADC value reaches the
host, which then linearizes. The forwarder already carries int16, and
both senders use sizeof(report), so this is the only change needed.

+++ include/zmk/hid-io/hid_joystick.h
@@ -15,12 +15,15 @@
 #include <zmk/hid-io/joystick.h>
 
 struct zmk_hid_joystick_report_body_alt {
-    int8_t d_x;
-    int8_t d_y;
-    int8_t d_z;
-    int8_t d_rx;
-    int8_t d_ry;
-    int8_t d_rz;
+    /* 16-bit axes: slide faders carry an absolute position whose useful
+     * resolution (esp. the taper-compressed top of travel) is far finer than
+     * 8 bits. X = left fader, Y = right fader. */
+    int16_t d_x;
+    int16_t d_y;
+    int16_t d_z;
+    int16_t d_rx;
+    int16_t d_ry;
+    int16_t d_rz;
     zmk_joystick_button_flags_t buttons;
 } __packed;
 struct zmk_hid_joystick_report_alt {
